Improved structure and reformatted code
