+++ GameHoarder/celery.py
@@ -1,5 +1,7 @@
 from __future__ import absolute_import
+
 import os
+
 from celery import Celery
 from django.conf import settings
