+++ gamehoarder_site/forms.py
@@ -1,9 +1,9 @@
-from django import forms
-from .models import *
 from django.contrib.auth.forms import UserCreationForm
+
+from .models import *
 
 
 class UserForm(UserCreationForm):
     class Meta:
         model = User
-        fields = ('first_name', 'last_name', 'username', 'email', 'password1', 'password2', )
+        fields = ('first_name', 'last_name', 'username', 'email', 'password1', 'password2',)
