+++ gamehoarder_site/templates/base_site.html
@@ -1,68 +1,68 @@
 <!DOCTYPE html>
 <html lang="en">
-  <head>
+<head>
     {% block head %}
-      {% block meta %}
-        <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-        <!-- Meta, title, CSS, favicons, etc. -->
-        <meta charset="utf-8">
-        <meta http-equiv="X-UA-Compatible" content="IE=edge">
-        <meta name="viewport" content="width=device-width, initial-scale=1">
-      {% endblock meta %}
+        {% block meta %}
+            <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
+            <!-- Meta, title, CSS, favicons, etc. -->
+            <meta charset="utf-8">
+            <meta http-equiv="X-UA-Compatible" content="IE=edge">
+            <meta name="viewport" content="width=device-width, initial-scale=1">
+        {% endblock meta %}
 
-      <title>GameHoarder | {% block title %}{% endblock %}</title>
+        <title>GameHoarder | {% block title %}{% endblock %}</title>
 
-      {% block stylesheets %}
-        <!-- Bootstrap -->
-        <link href="/static/vendors/bootstrap/dist/css/bootstrap.min.css" rel="stylesheet">
-        <!-- Font Awesome -->
-        <link href="/static/vendors/font-awesome/css/font-awesome.min.css" rel="stylesheet">
-        <!-- NProgress -->
-        <link href="/static/vendors/nprogress/nprogress.css" rel="stylesheet">
-        <!-- bootstrap-daterangepicker -->
-        <link href="/static/vendors/bootstrap-daterangepicker/daterangepicker.css" rel="stylesheet">
-        <!-- iCheck -->
-        <link href="/static/vendors/iCheck/skins/flat/green.css" rel="stylesheet">
-        <!-- bootstrap-progressbar -->
-        <link href="/static/vendors/bootstrap-progressbar/css/bootstrap-progressbar-3.3.4.min.css" rel="stylesheet">
-        <!-- JQVMap -->
-        <link href="/static/vendors/jqvmap/dist/jqvmap.min.css" rel="stylesheet"/>
-      {% endblock stylesheets %}
-      <!-- Custom Theme Style -->
-      <link href="/static/build/css/custom.css" rel="stylesheet">
+        {% block stylesheets %}
+            <!-- Bootstrap -->
+            <link href="/static/vendors/bootstrap/dist/css/bootstrap.min.css" rel="stylesheet">
+            <!-- Font Awesome -->
+            <link href="/static/vendors/font-awesome/css/font-awesome.min.css" rel="stylesheet">
+            <!-- NProgress -->
+            <link href="/static/vendors/nprogress/nprogress.css" rel="stylesheet">
+            <!-- bootstrap-daterangepicker -->
+            <link href="/static/vendors/bootstrap-daterangepicker/daterangepicker.css" rel="stylesheet">
+            <!-- iCheck -->
+            <link href="/static/vendors/iCheck/skins/flat/green.css" rel="stylesheet">
+            <!-- bootstrap-progressbar -->
+            <link href="/static/vendors/bootstrap-progressbar/css/bootstrap-progressbar-3.3.4.min.css" rel="stylesheet">
+            <!-- JQVMap -->
+            <link href="/static/vendors/jqvmap/dist/jqvmap.min.css" rel="stylesheet"/>
+        {% endblock stylesheets %}
+        <!-- Custom Theme Style -->
+        <link href="/static/build/css/custom.css" rel="stylesheet">
     {% endblock head %}
-  </head>
+</head>
 
-  <body class="{% block body_class %}nav-md {% endblock body_class %}">
-    {% block body %}
-      <div class="container body">
+<body class="{% block body_class %}nav-md {% endblock body_class %}">
+{% block body %}
+    <div class="container body">
         <div class="main_container">
 
-          {% block sidebar %}
-            <div class="col-md-3 left_col {% block sidebar_class %} {% endblock sidebar_class %}">
-              {% include "sidebar.html" %}
-            </div>
-          {% endblock sidebar %}
+            {% block sidebar %}
+                <div class="col-md-3 left_col {% block sidebar_class %} {% endblock sidebar_class %}">
+                    {% include "sidebar.html" %}
+                </div>
+            {% endblock sidebar %}
 
-          {% block top_navigation %}
-            <div class="top_nav">
-              {% include "top_navigation.html" %}
-            </div>
-          {% endblock top_navigation %}
+            {% block top_navigation %}
+                <div class="top_nav">
+                    {% include "top_navigation.html" %}
+                </div>
+            {% endblock top_navigation %}
 
-          {% block content %}
-            <!-- {{ content }} -->
-          {% endblock content %}
+            {% block content %}
+                <!-- {{ content }} -->
+            {% endblock content %}
 
-          {% block footer %}
-            <footer>
-              {% include "footer.html" %}
-            </footer>
-          {% endblock footer %}
+            {% block footer %}
+                <footer>
+                    {% include "footer.html" %}
+                </footer>
+            {% endblock footer %}
         </div>
-      </div>
+    </div>
 
-      {% block javascripts %}
+    {% block javascripts %}
         <!-- jQuery -->
         <script src="/static/vendors/jquery/dist/jquery.min.js"></script>
         <!-- Bootstrap -->
@@ -96,9 +96,9 @@
         <script src="/static/vendors/devbridge-autocomplete/dist/jquery.autocomplete.min.js"></script>
         <!-- starrr -->
         <script src="/static/vendors/starrr/dist/starrr.js"></script>
-      {% endblock %}
-      <!-- Custom Theme Scripts -->
-      <script src="/static/build/js/custom.js"></script>
-    {% endblock body %}
-  </body>
+    {% endblock %}
+    <!-- Custom Theme Scripts -->
+    <script src="/static/build/js/custom.js"></script>
+{% endblock body %}
+</body>
 </html>
