+++ gamehoarder_site/templates/collection/tables/finished_table.html
@@ -68,7 +68,7 @@
                                         </td>
                                         <td>
                                             <a href="/game/{{ title.game_version.db_id }}">
-                                            {{ title.game_version.parent_game.title }}</a>
+                                                {{ title.game_version.parent_game.title }}</a>
                                         </td>
                                         <td class="text-center">
                                             {% if title.time_played is not None %}
