+++ gamehoarder_site/templates/collection/tables/list_table.html
@@ -56,7 +56,7 @@
                                         </td>
                                         <td>
                                             <a href="/game/{{ title.game_version.db_id }}">
-                                            {{ title.game_version.parent_game.title }}</a>
+                                                {{ title.game_version.parent_game.title }}</a>
                                         </td>
                                         <td class="text-center">
                                             {{ title.game_version.parent_game.main_time }}
