+++ gamehoarder_site/templates/fixed_footer.html
@@ -3,27 +3,27 @@
 {% block title %} Fixed Footer {% endblock title %}
 
 {% block stylesheets %}
-  {{ block.super }}
-  <!-- jQuery custom content scroller -->
-  <link href="/static/vendors/malihu-custom-scrollbar-plugin/jquery.mCustomScrollbar.min.css" rel="stylesheet"/>
+    {{ block.super }}
+    <!-- jQuery custom content scroller -->
+    <link href="/static/vendors/malihu-custom-scrollbar-plugin/jquery.mCustomScrollbar.min.css" rel="stylesheet"/>
 {% endblock stylesheets %}
 
 {% block body_class %}nav-md footer_fixed{% endblock body_class %}
 
 {% block content %}
-  <div class="right_col" role="main">
-    <div class="">
-      <div class="page-title">
-        <div class="title_left">
-          <h3>Fixed Footer <small> Just add class <strong>footer_fixed</strong></small></h3>
+    <div class="right_col" role="main">
+        <div class="">
+            <div class="page-title">
+                <div class="title_left">
+                    <h3>Fixed Footer <small> Just add class <strong>footer_fixed</strong></small></h3>
+                </div>
+            </div>
         </div>
-      </div>
     </div>
-  </div>
 {% endblock content %}
 
 {% block javascripts %}
-  {{ block.super }}
-  <!-- jQuery custom content scroller -->
-  <script src="/static/vendors/malihu-custom-scrollbar-plugin/jquery.mCustomScrollbar.concat.min.js"></script>
+    {{ block.super }}
+    <!-- jQuery custom content scroller -->
+    <script src="/static/vendors/malihu-custom-scrollbar-plugin/jquery.mCustomScrollbar.concat.min.js"></script>
 {% endblock javascripts %}
