+++ gamehoarder_site/templates/sidebar.html
@@ -27,7 +27,8 @@
             <h3>{% trans "General" %}</h3>
             <ul class="nav side-menu">
                 <li><a href="{% url "index" %}"><i class="fa fa-home"></i> Home <span class="fa"></span></a></li>
-                <li><a href="{% url "friends" %}"><i class="fa fa-users"></i> {% trans "Friends" %} <span class="fa"></span></a></li>
+                <li><a href="{% url "friends" %}"><i class="fa fa-users"></i> {% trans "Friends" %} <span
+                        class="fa"></span></a></li>
             </ul>
         </div>
         <div class="menu_section">
@@ -51,10 +52,13 @@
         <div class="menu_section">
             <h3>{% trans "Lists" %}</h3>
             <ul class="nav side-menu">
-                <li><a href="{% url "interested_table" %}"><i class="fa fa-bookmark"></i> {% trans "Interested" %} <span class="fa"></span></a>
+                <li><a href="{% url "interested_table" %}"><i class="fa fa-bookmark"></i> {% trans "Interested" %} <span
+                        class="fa"></span></a>
                 </li>
-                <li><a href="{% url "wishlist_table" %}"><i class="fa fa-star"></i> {% trans "Wishlist" %} <span class="fa"></span></a></li>
-                <li><a><i class="fa fa-institution"></i> {% trans "Custom" %} <span class="fa fa-chevron-down"></span></a>
+                <li><a href="{% url "wishlist_table" %}"><i class="fa fa-star"></i> {% trans "Wishlist" %} <span
+                        class="fa"></span></a></li>
+                <li><a><i class="fa fa-institution"></i> {% trans "Custom" %} <span
+                        class="fa fa-chevron-down"></span></a>
                     <ul class="nav child_menu">
                         {% for tag in tags %}
                             <li><a href="/tag/?tag={{ tag.name }}">{{ tag.name }}</a></li>
@@ -67,10 +71,10 @@
     <!-- /sidebar menu -->
 
     <!-- /menu footer buttons -->
-        <div class="sidebar-footer hidden-small">
-          <a data-toggle="tooltip" data-placement="top" title="Logout" href="/logout">
+    <div class="sidebar-footer hidden-small">
+        <a data-toggle="tooltip" data-placement="top" title="Logout" href="/logout">
             <span class="glyphicon glyphicon-off" aria-hidden="true"></span>
-          </a>
-        </div>
+        </a>
+    </div>
     <!-- /menu footer buttons -->
 </div>
