+++ gamehoarder_site/templates/top_navigation.html
@@ -1,92 +1,93 @@
 <div class="nav_menu">
-  <nav>
-    <div class="nav toggle">
-      <a id="menu_toggle"><i class="fa fa-bars"></i></a>
-    </div>
+    <nav>
+        <div class="nav toggle">
+            <a id="menu_toggle"><i class="fa fa-bars"></i></a>
+        </div>
 
-    <ul class="nav navbar-nav navbar-right">
-      <li class="">
-        <a href="javascript:;" class="user-profile dropdown-toggle" data-toggle="dropdown" aria-expanded="false">
-          <img src="/static/images/img.jpg" alt="">John Doe
-          <span class=" fa fa-angle-down"></span>
-        </a>
-        <ul class="dropdown-menu dropdown-usermenu pull-right">
-          <li><a href="javascript:;"> Profile</a></li>
-          <li>
-            <a href="javascript:;">
-              <span class="badge bg-red pull-right">50%</span>
-              <span>Settings</span>
-            </a>
-          </li>
-          <li><a href="javascript:;">Help</a></li>
-          <li><a href="login.html"><i class="fa fa-sign-out pull-right"></i> Log Out</a></li>
-        </ul>
-      </li>
+        <ul class="nav navbar-nav navbar-right">
+            <li class="">
+                <a href="javascript:" class="user-profile dropdown-toggle" data-toggle="dropdown"
+                   aria-expanded="false">
+                    <img src="/static/images/img.jpg" alt="">John Doe
+                    <span class=" fa fa-angle-down"></span>
+                </a>
+                <ul class="dropdown-menu dropdown-usermenu pull-right">
+                    <li><a href="javascript:"> Profile</a></li>
+                    <li>
+                        <a href="javascript:">
+                            <span class="badge bg-red pull-right">50%</span>
+                            <span>Settings</span>
+                        </a>
+                    </li>
+                    <li><a href="javascript:">Help</a></li>
+                    <li><a href="login.html"><i class="fa fa-sign-out pull-right"></i> Log Out</a></li>
+                </ul>
+            </li>
 
-      <li role="presentation" class="dropdown">
-        <a href="javascript:;" class="dropdown-toggle info-number" data-toggle="dropdown" aria-expanded="false">
-          <i class="fa fa-envelope-o"></i>
-          <span class="badge bg-green">6</span>
-        </a>
-        <ul id="menu1" class="dropdown-menu list-unstyled msg_list" role="menu">
-          <li>
-            <a>
-              <span class="image"><img src="/static/images/img.jpg" alt="Profile Image" /></span>
-              <span>
+            <li role="presentation" class="dropdown">
+                <a href="javascript:" class="dropdown-toggle info-number" data-toggle="dropdown" aria-expanded="false">
+                    <i class="fa fa-envelope-o"></i>
+                    <span class="badge bg-green">6</span>
+                </a>
+                <ul id="menu1" class="dropdown-menu list-unstyled msg_list" role="menu">
+                    <li>
+                        <a>
+                            <span class="image"><img src="/static/images/img.jpg" alt="Profile Image"/></span>
+                            <span>
                 <span>John Doe</span>
                 <span class="time">3 mins ago</span>
               </span>
-              <span class="message">
+                            <span class="message">
                 Film festivals used to be do-or-die moments for movie makers. They were where...
               </span>
-            </a>
-          </li>
-          <li>
-            <a>
-              <span class="image"><img src="/static/images/img.jpg" alt="Profile Image" /></span>
-              <span>
+                        </a>
+                    </li>
+                    <li>
+                        <a>
+                            <span class="image"><img src="/static/images/img.jpg" alt="Profile Image"/></span>
+                            <span>
                 <span>John Doe</span>
                 <span class="time">3 mins ago</span>
               </span>
-              <span class="message">
+                            <span class="message">
                 Film festivals used to be do-or-die moments for movie makers. They were where...
               </span>
-            </a>
-          </li>
-          <li>
-            <a>
-              <span class="image"><img src="/static/images/img.jpg" alt="Profile Image" /></span>
-              <span>
+                        </a>
+                    </li>
+                    <li>
+                        <a>
+                            <span class="image"><img src="/static/images/img.jpg" alt="Profile Image"/></span>
+                            <span>
                 <span>John Doe</span>
                 <span class="time">3 mins ago</span>
               </span>
-              <span class="message">
+                            <span class="message">
                 Film festivals used to be do-or-die moments for movie makers. They were where...
               </span>
-            </a>
-          </li>
-          <li>
-            <a>
-              <span class="image"><img src="/static/images/img.jpg" alt="Profile Image" /></span>
-              <span>
+                        </a>
+                    </li>
+                    <li>
+                        <a>
+                            <span class="image"><img src="/static/images/img.jpg" alt="Profile Image"/></span>
+                            <span>
                 <span>John Doe</span>
                 <span class="time">3 mins ago</span>
               </span>
-              <span class="message">
+                            <span class="message">
                 Film festivals used to be do-or-die moments for movie makers. They were where...
               </span>
-            </a>
-          </li>
-          <li>
-            <div class="text-center">
-              <a>
-                <strong>See All Alerts</strong>
-                <i class="fa fa-angle-right"></i>
-              </a>
-            </div>
-          </li>
+                        </a>
+                    </li>
+                    <li>
+                        <div class="text-center">
+                            <a>
+                                <strong>See All Alerts</strong>
+                                <i class="fa fa-angle-right"></i>
+                            </a>
+                        </div>
+                    </li>
+                </ul>
+            </li>
         </ul>
-      </li>
-    </ul>
-  </nav>
+    </nav>
 </div>
