+++ gamehoarder_site/templates/upload.html
@@ -10,7 +10,6 @@
 <body>
 
 
-
 </body>
 {% if task_id %}
     <script src="{% static 'celery_progress/celery_progress.js' %}"></script>
