+++ gamehoarder_site/urls.py
@@ -4,7 +4,7 @@ from . import views
 
 urlpatterns = [
     path('', views.index, name='index'),
-     
+
     path('login', views.login_register, name='login'),
     path('logout', views.logout, name='logout'),
     path('friends', views.friends, name='friends'),
