+++ gamehoarder_site/views.py
@@ -53,6 +53,7 @@ def download_csv(request):
 
         return response
 
+
 def login_register(request):
     if request.method == 'POST':
         form = UserForm(request.POST)
